requirements.txt
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.0
colorama==0.4.6
Django==5.1.4
django-extensions==3.2.3
djangorestframework==3.15.2
idna==3.10
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
pytest==8.3.4
pytest-django==4.9.0
pytest-mock==3.14.0
requests==2.32.3
sqlparse==0.5.3
tzdata==2024.2
urllib3==2.2.3
django-environ==0.10.0
gunicorn==20.1.0
psycopg2-binary==2.9.9